simple restructuring
