+++ db_setup.js
@@ -1,0 +1,74 @@
+//
+//  Copyright (c) 2018, Respective Authors all rights reserved.
+//
+//  The MIT License
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to
+//  deal in the Software without restriction, including without limitation the
+//  rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+//  sell copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+//  IN THE SOFTWARE.
+//
+
+const SQLITE_FILENAME = process.env.SQLITE_FILENAME || "database.sqlite";
+
+const Sequelize = require('sequelize');
+
+const makeSequelize = function () {
+    if (process.env.ENV_PRODUCTION && process.env.LUMEOS_SERVER_DB) {
+        console.log("Connecting with: " + process.env.LUMEOS_SERVER_DB);
+        return new Sequelize(process.env.LUMEOS_SERVER_DB, {
+            operatorsAliases: false,
+            pool: {
+                max: 5,
+                min: 0,
+                acquire: 30000,
+                idle: 10000
+            }
+        });
+    } else if (process.env.ENV_PRODUCTION) {
+        throw "Production env is specified, but LUMEOS_SERVER_DB is not set";
+    }
+
+    return new Sequelize('database', 'username', 'password', {
+        host: 'localhost',
+        dialect: 'sqlite',
+
+        operatorsAliases: false,
+        pool: {
+            max: 5,
+            min: 0,
+            acquire: 30000,
+            idle: 10000
+        },
+
+        storage: SQLITE_FILENAME
+    });
+}
+
+const sequelize = makeSequelize();
+
+sequelize
+    .authenticate()
+    .then(() => {
+        console.log('Connection has been established successfully.');
+    })
+    .catch(err => {
+        console.error('Unable to connect to the database:', err);
+    });
+
+module.exports = {
+    dbInstance: sequelize
+}
